LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/unreachable-snap-positions-expected.txt


FAIL Snaps to the positions defined by the element as much as possible assert_equals: expected 0 but got 100
